requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
notion-client==2.2.1
python-dotenv==1.0.1
pydantic>=2.0.0
qrcode==7.4.2
Pillow==10.2.0


